Add files via upload
The ATMs in our cities are built on Java, as we have all seen them. It is a console-based application with five different classes. To use the system, the user must enter his or her user ID and PIN when it starts. Once the details are entered successfully, ATM functionality is unlocked. As a result of the project, the following operations can be performed:

Transactions History
Withdraw
Deposit
Transfer
Quit

Here's a basic implementation of an ATM system in Java. This implementation includes five classes: ATM, User, Transaction, Bank, and Main. The system allows a user to log in and once authenticated, they can perform various operations such as viewing transaction history, withdrawing money, depositing money, transferring money, and quitting the application.